+++ .factory/automation.yml
@@ -55,7 +55,7 @@ build:
         bazel run @vaticle_dependencies//distribution/artifact:create-netrc
         bazel build //...
         tools/start-core-server.sh
-        bazel test //tests:queries --test_arg=-- --test_arg=core --test_arg=--test-threads=1 --test_output=streamed && export TEST_SUCCESS=0 || export TEST_SUCCESS=1
+        bazel test //tests --test_arg=-- --test_arg=integration::queries::core --test_arg=--test-threads=1 --test_output=streamed && export TEST_SUCCESS=0 || export TEST_SUCCESS=1
         tools/stop-core-server.sh
         exit $TEST_SUCCESS
     test-integration-cluster:
@@ -67,7 +67,7 @@ build:
         bazel run @vaticle_dependencies//distribution/artifact:create-netrc
         bazel build //...
         source tools/start-cluster-servers.sh # use source to receive export vars
-        bazel test //tests:queries --test_env=ROOT_CA=$ROOT_CA --test_arg=-- --test_arg=cluster --test_arg=--test-threads=1 --test_output=streamed && export TEST_SUCCESS=0 || export TEST_SUCCESS=1
+        bazel test //tests --test_env=ROOT_CA=$ROOT_CA --test_arg=-- --test_arg=integration::queries::cluster --test_arg=--test-threads=1 --test_output=streamed && export TEST_SUCCESS=0 || export TEST_SUCCESS=1
         tools/stop-cluster-servers.sh
         exit $TEST_SUCCESS
     test-integration-runtimes:
@@ -78,7 +78,7 @@ build:
         bazel run @vaticle_dependencies//distribution/artifact:create-netrc
         bazel build //...
         tools/start-core-server.sh
-        bazel test //tests:runtimes --test_arg=-- --test_arg=--test-threads=1 --test_output=streamed && export TEST_SUCCESS=0 || export TEST_SUCCESS=1
+        bazel test //tests --test_arg=-- --test_arg=integration::runtimes --test_arg=--test-threads=1 --test_output=streamed && export TEST_SUCCESS=0 || export TEST_SUCCESS=1
         tools/stop-core-server.sh
         exit $TEST_SUCCESS
     deploy-crate-snapshot:
