+++ tests/behaviour/connection/database/steps.rs
@@ -25,7 +25,10 @@ use cucumber::{gherkin::Step, given, then, when};
 use futures::{future::try_join_all, TryFutureExt};
 use typedb_client::Database;
 
-use crate::{generic_step_impl, steps::util, TypeDBWorld};
+use crate::{
+    behaviour::{util, TypeDBWorld},
+    generic_step_impl,
+};
 
 generic_step_impl! {
     #[step(expr = "connection create database: {word}")]
