+++ tests/integration/runtimes.rs
@@ -19,11 +19,11 @@
  * under the License.
  */
 
-mod common;
-
 use futures::StreamExt;
 use serial_test::serial;
 use typedb_client::{DatabaseManager, Session, SessionType::Data, TransactionType::Write};
+
+use super::common;
 
 #[test]
 #[serial]
